Move send_tcp_v6
